requirements.txt
fastapi
uvicorn
python-multipart
groq
openai-whisper
PyPDF2
pinecone-client
langchain-pinecone
langchain-huggingface
sentence-transformers
gtts
upstash-redis
python-dotenv
flashrank